kotlinx-metadata-jvm: always use latest metadata/bytecode versions in class writers
The main reason for this change is that the current API for class
version requirements (`KmClass.versionRequirements`) makes it impossible
to support pre-1.4 metadata where this was stored incorrectly for nested
classes: with the "version requirement table" in the outer class, and
indexes into that table in nested classes. See KT-41011.

Other than this aspect, metadata of classes is basically the same in
pre-1.4 and 1.4, which means that if some kotlinx-metadata-jvm users
really need to serialize metadata of an earlier version and they don't
care about the version requirements issue, they can just use these new
bytes but write the earlier version (e.g. 1.1) to the class file.
Everything will work the same, except for the possible version
requirements issue mentioned above.

Note that we still provide a way to specify custom metadata version for
`KotlinModuleMetadata.Writer` though, since kotlinx-metadata-jvm
supports writing both pre-1.4 and 1.4 .kotlin_module files, and their
data layout is completely different, so the version actually matters.

Also, do not provide a way to write custom bytecode version. It has
never been used meaningfully in Kotlin, and will be removed in a future
release, so there's no point in using it. See KT-41758.

 #KT-41011 Fixed